goal routes authenticated and controller tailored for authorized users

+++ Week 6/Day 1/backend/controllers/goalController.js
@@ -1,10 +1,11 @@
 const asyncHandler = require('express-async-handler')
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
 // @desc: Get Goals
 // @route: GET /api/goals
 // @access: Private
 const getGoals = asyncHandler( async (req, res) => { //using async, before getting into the database, becuase it will take some time to get the data
-    const goals = await Goal.find({}) //find all the goals    
+    const goals = await Goal.find({ user: req.user.id }) //find all the goals    
         res.status(200).json(goals) //send the goals in json format
 })
 
@@ -16,7 +17,10 @@ const postGoals = asyncHandler (async (req, res) => {
         res.status(400)
         throw new Error('Please add text field')
     }
-    const goal = new Goal({text: req.body.text})
+    const goal = new Goal({
+        text: req.body.text,
+        user: req.user.id
+    })
     const createdGoal = await goal.save()
     res.status(200).json(createdGoal)
 })
@@ -29,6 +33,20 @@ const putGoals = asyncHandler (async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if (!user) {
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    // Make sure user is goal owner, logged in user matches the user who created the goal (goal user)
+    if (goal.user.toString() !== req.user.id) { // goal.user is an object, so we need to convert it to string, why we are using goal.user and how? , because we have user field in goal model
+        res.status(401)
+        throw new Error('User not authorized')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -48,8 +66,24 @@ const deleteGoals = asyncHandler (async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
-    // await goal.destroy() // not working, becuse we are using mongoose version above 5
 
+    
+    // Get user
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if (!user) {
+        res.status(400)
+        throw new Error('User not found')
+    }
+
+    // Make sure user is goal owner, logged in user matches the user who created the goal (goal user)
+    if (goal.user.toString() !== req.user.id) { // goal.user is an object, so we need to convert it to string, why we are using goal.user and how? , because we have user field in goal model
+        res.status(401)
+        throw new Error('User not authorized')
+    }
+
+    // await goal.destroy() // not working, becuse we are using mongoose version above 5
     await goal.deleteOne();
     
     res.status(200).json({message: 'Goal removed', id: req.params.id})
